requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Vector Database & Embeddings
chromadb>=0.4.18
sentence-transformers>=2.2.2

# LangChain & LLM
langchain-groq>=0.1.0
langchain-core>=0.1.0

# Database
pymongo>=4.6.0

# Environment Variables
python-dotenv>=1.0.0
